fix: use virtual columns
Previously, whenever a DBMS inserted tab literals ('\t') into the
`dbout`, it would mess up the cell range selection. Now, we report back
two sets of numbers:

- The "real" {to, from} which describes the column range on the current
  line.
- The "virtual" {to, from} which describes the column range as if there
  were no tabs on the line.

There are still some issues to overcome (such as the query not
generating the correct results when run automatically, but when run
manually the expected result occurs.)

However, this is a good start.