refactor: set `filetype` after other options
This allows filetype plugins, such as `mysql.vim`, to override any of
the special query buffer settings if needed.

A similar pattern of setting `filetype` before other settings is present
for the DBUI drawer. I assume this is more deliberate, and is out of
scope for kristijanhusak/vim-dadbod-ui#267, so I haven't changed this
part.

This change was not signed off, so might be reverted.

BREAKING-CHANGE: Options set in filetype plugins are now preferred for
		 query buffers.
Link: https://github.com/kristijanhusak/vim-dadbod-ui/blob/0f51d8de368c8c6220973e8acd156d17da746f4c/autoload/db_ui/drawer.vim?plain=1#L38